+++ app/src/main/java/com/programmers/java/chap1/item1/builder/domain/Calzone.java
@@ -1,0 +1,33 @@
+package com.programmers.java.chap1.item1.builder.domain;
+
+import com.programmers.java.chap1.item1.builder.domain._abstract.Pizza;
+
+public class Calzone extends Pizza {
+
+    private final boolean sauceInside;
+
+    public static class Builder extends Pizza.Builder<Builder> {
+
+        private boolean sauceInside = false;
+
+        public Builder sauceInside() {
+            sauceInside = true;
+            return this;
+        }
+
+        @Override
+        public Calzone build() {
+            return new Calzone(this);
+        }
+
+        @Override
+        protected Builder self() {
+            return this;
+        }
+    }
+
+    private Calzone(Builder builder) {
+        super(builder);
+        sauceInside = builder.sauceInside;
+    }
+}
