+++ app/src/main/java/com/programmers/java/chap1/item1/builder/domain/NyPizza.java
@@ -1,0 +1,36 @@
+package com.programmers.java.chap1.item1.builder.domain;
+
+import com.programmers.java.chap1.item1.builder.domain._abstract.Pizza;
+import java.util.Objects;
+
+public class NyPizza extends Pizza {
+
+    public enum Size {SMALL, MEDIUM, LARGE}
+
+    private final Size size;
+
+    public static class Builder extends Pizza.Builder<Builder> {
+
+        private final Size size;
+
+        public Builder(Size size) {
+            this.size = Objects.requireNonNull(size);
+        }
+
+        @Override
+        public NyPizza build() {
+            return new NyPizza(this);
+        }
+
+        @Override
+        protected Builder self() {
+            return this;
+        }
+    }
+
+    private NyPizza(Builder builder) {
+        super(builder);
+        size = builder.size;
+    }
+
+}
