+++ app/src/main/java/com/programmers/java/chap1/item1/singleton/Main.java
@@ -1,0 +1,5 @@
+package com.programmers.java.chap1.item1.singleton;
+
+public class Main {
+
+}
