+++ app/src/main/java/com/programmers/java/chap1/item1/singleton/domain/EnumTypeSingleton.java
@@ -1,0 +1,9 @@
+package com.programmers.java.chap1.item1.singleton.domain;
+
+public enum EnumTypeSingleton {
+    INSTANCE;
+
+    public void test(){
+        System.out.println("test");
+    }
+}
