+++ app/src/main/java/com/programmers/java/chap1/item4/domain/A.java
@@ -1,0 +1,5 @@
+package com.programmers.java.chap1.item4.domain;
+
+public class A extends TypeFactory {
+
+}
