+++ app/src/main/java/com/programmers/java/chap1/item4/domain/factorymethod/Tea.java
@@ -1,0 +1,9 @@
+package com.programmers.java.chap1.item4.domain.factorymethod;
+
+public interface Tea {
+
+    String drink();
+
+    int teaType();
+
+}
