+++ app/src/main/java/com/programmers/java/chap1/item8/Adult.java
@@ -1,0 +1,9 @@
+package com.programmers.java.chap1.item8;
+
+public class Adult {
+	public static void main(String[] args) {
+		try (Room myRoom=new Room(7)){
+			System.out.println("hello !");
+		}
+	}
+}
