+++ app/src/main/java/com/programmers/java/chap1/item8/Main.java
@@ -1,0 +1,4 @@
+package com.programmers.java.chap1.item8;
+
+public class Main {
+}
