--- app/src/main/java/com/programmers/java/item1/Dummy.java
@@ -1,5 +1,0 @@
-package com.programmers.java.item1;
-
-public class Dummy {
-
-}
