--- app/src/main/java/com/programmers/java/item1/code/yeon/builder/domain/Calzone.java
@@ -1,33 +1,0 @@
-package com.programmers.java.item1.code.yeon.builder.domain;
-
-import com.programmers.java.item1.code.yeon.builder.domain._abstract.Pizza;
-
-public class Calzone extends Pizza {
-
-    private final boolean sauceInside;
-
-    public static class Builder extends Pizza.Builder<Builder> {
-
-        private boolean sauceInside = false;
-
-        public Builder sauceInside() {
-            sauceInside = true;
-            return this;
-        }
-
-        @Override
-        public Calzone build() {
-            return new Calzone(this);
-        }
-
-        @Override
-        protected Builder self() {
-            return this;
-        }
-    }
-
-    private Calzone(Builder builder) {
-        super(builder);
-        sauceInside = builder.sauceInside;
-    }
-}
