--- app/src/main/java/com/programmers/java/item1/code/yeon/builder/domain/NutritionFacts.java
@@ -1,68 +1,0 @@
-package com.programmers.java.item1.code.yeon.builder.domain;
-
-public class NutritionFacts {
-    private final int servingSize;
-    private final int servings;
-    private final int calories;
-    private final int fat;
-    private final int sodium;
-    private final int carbohydrate;
-
-    private NutritionFacts(Builder builder) {
-        servingSize = builder.servingSize;
-        servings = builder.servings;
-        calories = builder.calories;
-        fat = builder.fat;
-        sodium = builder.sodium;
-        carbohydrate = builder.carbohydrate;
-    }
-
-    public static void loadTest(){
-        System.out.println("load test");
-    }
-
-    public static class Builder {
-        // NOTE : 필수 매개변수
-        private final int servingSize;
-        private final int servings;
-
-        // NOTE : 선택 매개변수
-        private int calories = 0;
-        private int fat = 0;
-        private int sodium = 0;
-        private int carbohydrate = 0;
-
-        // NOTE : 필수 매개변수만을 담은 Builder 생성자
-        public Builder(int servingSize, int servings) {
-            this.servingSize = servingSize;
-            this.servings = servings;
-        }
-
-        // 선택 매개변수의 setter, Builder 자신을 반환해 연쇄적으로 호출 가능
-        public Builder calories(int val) {
-            calories = val;
-            return this;
-        }
-
-        public Builder fat(int val) {
-            fat = val;
-            return this;
-        }
-
-        public Builder sodium(int val) {
-            sodium = val;
-            return this;
-        }
-
-        public Builder carbohydrate(int val) {
-            carbohydrate = val;
-            return this;
-        }
-
-        // NOTE : build() 최종 불변 객체를 얻는다.
-        public NutritionFacts build() {
-            return new NutritionFacts(this);
-        }
-    }
-
-}
