--- app/src/main/java/com/programmers/java/item1/code/yeon/builder/domain/NyPizza.java
@@ -1,36 +1,0 @@
-package com.programmers.java.item1.code.yeon.builder.domain;
-
-import com.programmers.java.item1.code.yeon.builder.domain._abstract.Pizza;
-import java.util.Objects;
-
-public class NyPizza extends Pizza {
-
-    public enum Size {SMALL, MEDIUM, LARGE}
-
-    private final Size size;
-
-    public static class Builder extends Pizza.Builder<Builder> {
-
-        private final Size size;
-
-        public Builder(Size size) {
-            this.size = Objects.requireNonNull(size);
-        }
-
-        @Override
-        public NyPizza build() {
-            return new NyPizza(this);
-        }
-
-        @Override
-        protected Builder self() {
-            return this;
-        }
-    }
-
-    private NyPizza(Builder builder) {
-        super(builder);
-        size = builder.size;
-    }
-
-}
