--- app/src/main/java/com/programmers/java/item1/code/yeon/singleton/Main.java
@@ -1,5 +1,0 @@
-package com.programmers.java.item1.code.yeon.singleton;
-
-public class Main {
-
-}
