--- app/src/main/java/com/programmers/java/item1/code/yeon/singleton/domain/EagerSingleton.java
@@ -1,15 +1,0 @@
-package com.programmers.java.item1.code.yeon.singleton.domain;
-
-public class EagerSingleton {
-
-    public static final EagerSingleton instance = new EagerSingleton();
-
-    private EagerSingleton() {
-
-    }
-
-    public void test() {
-        System.out.println("test");
-    }
-
-}
