--- app/src/main/java/com/programmers/java/item1/code/yeon/singleton/domain/EnumTypeSingleton.java
@@ -1,9 +1,0 @@
-package com.programmers.java.item1.code.yeon.singleton.domain;
-
-public enum EnumTypeSingleton {
-    INSTANCE;
-
-    public void test(){
-        System.out.println("test");
-    }
-}
