--- app/src/main/java/com/programmers/java/item4/code/yeon/domain/A.java
@@ -1,5 +1,0 @@
-package com.programmers.java.item4.code.yeon.domain;
-
-public class A extends TypeFactory {
-
-}
