--- app/src/main/java/com/programmers/java/item4/code/yeon/domain/B.java
@@ -1,5 +1,0 @@
-package com.programmers.java.item4.code.yeon.domain;
-
-public class B extends TypeFactory {
-
-}
