--- app/src/main/java/com/programmers/java/item4/code/yeon/domain/factorymethod/Programmer.java
@@ -1,13 +1,0 @@
-package com.programmers.java.item4.code.yeon.domain.factorymethod;
-
-public class Programmer {
-    private Tea tea;
-
-    public Programmer(Tea tea) {
-        this.tea = tea;
-    }
-
-    public void drink(){
-        System.out.println(tea.drink());
-    }
-}
