--- app/src/main/java/com/programmers/java/item4/code/yeon/domain/factorymethod/Tea.java
@@ -1,9 +1,0 @@
-package com.programmers.java.item4.code.yeon.domain.factorymethod;
-
-public interface Tea {
-
-    String drink();
-
-    int teaType();
-
-}
