--- app/src/main/java/com/programmers/java/item8/code/yeon/Main.java
@@ -1,4 +1,0 @@
-package com.programmers.java.item8.code.yeon;
-
-public class Main {
-}
